requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
requests
python-dateutil